Bug management

+++ core/headers/Point.h
@@ -21,7 +21,7 @@ public:
 		bool operator!=(const Point &p) const;
 
 		//typedef long double IndexType;
-		typedef unsigned int IndexType;
+		typedef unsigned long int IndexType;
 
     float x, y, z;
 		IndexType index;
